Backend/requirements.txt
astroid==2.4.2
azure-cognitiveservices-vision-computervision==0.5.0
azure-common==1.1.25
azure-core==1.6.0
azure-functions==1.2.1
azure-storage-file-share==12.1.2
certifi==2020.4.5.2
cffi==1.14.0
chardet==3.0.4
click==7.1.2
colorama==0.4.3
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.9.2
cycler==0.10.0
dataloader==2.0
docopt==0.6.2
editdistance==0.5.3
Flask==1.1.2
idna==2.9
inexactsearch==1.0.2
isodate==0.6.0
isort==4.3.21
itsdangerous==1.1.0
itypes==1.2.0
Jinja2==2.11.2
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
matplotlib==3.2.1
mccabe==0.6.1
msrest==0.6.16
nltk==3.2.4
numpy==1.18.2
oauthlib==3.1.0
openapi-codec==1.3.2
opencv-contrib-python==4.2.0.34
opencv-python==4.2.0.34
pandas==1.0.3
Pillow==7.1.2
pipreqs==0.4.10
preprocessing==0.1.13
preprocessor==1.1.3
ptvsd==4.3.2
pycparser==2.20
pyenchant==3.1.1
pylint==2.5.3
pyparsing==2.4.7
pyspellchecker==0.5.4
python-dateutil==2.8.1
pytz==2020.1
requests==2.23.0
requests-oauthlib==1.3.0
silpa-common==0.3
six==1.14.0
soundex==1.1.3
spellchecker==0.4
sphinx-rtd-theme==0.2.4
toml==0.10.1
uritemplate==3.0.1
urllib3==1.25.9
Werkzeug==1.0.1
wrapt==1.12.1
yarg==0.1.9
